[Silver III] Title: 교수가 된 현우, Time: 16 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/3474. 교수가 된 현우/교수가 된 현우.cc
@@ -9,15 +9,14 @@ int main() {
 	int T;
 	cin >> T;
 	while (T--) {
-		long long int n;
+		int n;
 		cin >> n;
 		int cnt = 0;
-		long long int k = 5;
+		int k = 5;
 		while (true) {
 			if (k > n) break;
 			cnt += (n / k);
 			k *= 5;
-
 		}
 		cout << cnt << '\n';
 	}
